# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prakash Knowledge Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  content: string;
  /** @nullable */
  featuredImage?: string | null;
  category: string;
  tags?: string[];
  externalLinks?: string[];
  isFeatured?: boolean;
  publishedAt: string;
  readingTime: number;
  viewCount: number;
  likeCount: number;
}

export interface PostList {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface PostInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  /** @minLength 1 */
  content: string;
  /** @nullable */
  featuredImage?: string | null;
  category: string;
  tags?: string[];
  externalLinks?: string[];
  isFeatured?: boolean;
}

export interface PostUpdate {
  title?: string;
  /** @nullable */
  shortDescription?: string | null;
  content?: string;
  /** @nullable */
  featuredImage?: string | null;
  category?: string;
  tags?: string[];
  externalLinks?: string[];
  isFeatured?: boolean;
}

export interface Category {
  name: string;
  slug: string;
  postCount: number;
}

export interface Tag {
  name: string;
  slug: string;
  postCount: number;
}

export interface ViewCount {
  viewCount: number;
}

export interface LikeCount {
  likeCount: number;
  liked: boolean;
}

export interface SiteSummary {
  totalPosts: number;
  totalViews: number;
  totalLikes: number;
  categoryCount: number;
}

export type ListPostsParams = {
category?: string;
tag?: string;
page?: number;
limit?: number;
};

export type ListPopularPostsParams = {
limit?: number;
};

export type ListRecentPostsParams = {
limit?: number;
};

export type SearchPostsParams = {
q: string;
limit?: number;
};