# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prakash Knowledge Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  name: string;
  slug: string;
  postCount: number;
}