# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prakash Knowledge Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  content: string;
  /** @nullable */
  featuredImage?: string | null;
  category: string;
  tags?: string[];
  externalLinks?: string[];
  isFeatured?: boolean;
  publishedAt: Date;
  readingTime: number;
  viewCount: number;
  likeCount: number;
}