# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prakash Knowledge Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  /** @minLength 1 */
  content: string;
  /** @nullable */
  featuredImage?: string | null;
  category: string;
  tags?: string[];
  externalLinks?: string[];
  isFeatured?: boolean;
}