# lib/api-zod/src/generated/types/postList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prakash Knowledge Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface PostList {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}