# lib/api-zod/src/generated/types/siteSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prakash Knowledge Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSummary {
  totalPosts: number;
  totalViews: number;
  totalLikes: number;
  categoryCount: number;
}